Use an earlier hook in the constructor
The idea behind a low (negative) priority is to make sure you get in befor any other code tries to use this functionality.  That seems like a good use for a hook earlier than init to me.

'plugins_loaded' is a good place to count on this code being loaded before others try to use it.