- updated pruning get_posts args and docs
This updates the get_posts args to 100 posts. It would have just grabbed
the default for the site which is 10 posts. Few sites that would take
the time to log would ever catch up to generated logs if they only
pruned 10 logs.

Now we prune 100 logs every time the routines are run.

Docs are update to recommend hourly pruning and add details of the
get_posts args changes.